Changed file list in specfile
Previously, the file list was determined by scanning through the source files
with a find(1)-based pipeline. This approach doesn't work when rpmbuild(1)
compiles Python files after when the scan is run. In any case, apart from
compiled Python files, the only files in the target directory are ones we
explicitly copy over, so this commit cuts out the scan.

In addition, this commit reduces some duplication in the %install section, and
marks the doc files as such.