Adding a value for serviceType, and a fully defined ScheduleType
Although the nsi spec states that startTime and endTime are optional
params for ScheduleType, opennsa expects these to be set when querying
for a connection. startTime has been set to 2017-03-13 (the day I
found this issue), and endTime has been set to 2222-02-22. These are
placeholder values only and should not be relied upon for circuits
that were created without specifying startTime or endTime.