get_of_circuit_data: fixed null-interface issue; refactored logic ISSUE=4238
When one tried to use get_of_circuit data without specifying an interface name,
it would try to use the interface name for the first endpoint... even when
one was querying for a different node. In addition, if the interface specified
(or selected by the above procedure and mis-applied to the node in question)
wasn't part of the active path for the circuit in question, then the code
would blithely use an interface of undef, leading to unsuable results.

This commit includes fixes for both the above issues. In addition, it refactors
some variable and field names to hopefully be a little less confusing (e.g.,
'node' now always refers to node names and never to DPIDs), and removes a
hash that wasn't strictly necessary for the operation of the method.