Multi-link popup panel: remove from DOM when closing, not just hide ISSUE=4527
Simply hiding the panel causes problems when we want to use a multi-link panel
again, as the way the app does things, it generates a new <div> with the exact
same id. Duplicate nodes with the same ID lead to the panel not working (to the
point of not displaying, even), at least on Firefox.

This change has been tested with Firefox and Chrome.