Updating handle_link to use new isis struct
_process_adjacencies was overriding some adjacencies when multiple
links existed between two nodes. We now reurn a hash containing all
adjacencies from this function, and use the ipv4_intf map in
Discovery.pm to resolve the remote interface associated with each
adjacency. With the remote interface, we can lookup the other end of
the isis adjacency; This gives us all the information we need to
populate the database.