Using mgmt_address as the dpid for netconf only nodes
For NETCONF only nodes we intended to store the integer representation
of the mgmt_address as the dpid, but the code was actually storing a
packed structure. This might not have been an issue, except that the
JSON perl module couldn't always decode these values. Calling unpack
on this strucutre is enough to get the integer value we wanted, and
fixes the json decoding problem.