Re-adding diffing on strict paths
Diffing on strict paths was broken; If a mpls path on a device was
modified OESS was unable to repair the change. This change allows OESS
to detect extra, missing, and incorrect hops and then correct them.

This change was removed prior to the last release. It's now being
pulled back in to fix incorrectly defined static paths, and migrate to
loopback based link discovery.