Updating how errors are handled in _edit_vlan and commit
Adding error handling for a case in _edit_vlan where error messages
come back in a slightly different form.

Some fatal errors aren't reported by the Junipers' until the config is
committed. We added extra error handling in commit to ensure the
correct result was returned, and added a check for this result in
add_vlan.