Fixing error check that caused path instantiation entries to corrupt
A db update that is successful, but updates no rows will return
0. This causes truthy error checks to fail. This change updates our
error checks on update in the _change_active_path method.

It was also discovered that path instantiations were not being created
for both paths when a failover occurred. We now set the end_epoch on
both path's existing path_instantiation entries. We also create a new
path_instantiation entry for each path with an end_epoch of -1.