Added duplicate-entry elimination in OESS::DB::Entity methods #531
If there are multiple instances of the same {user, interface, parent, child}
in the OESS::Entity object, previously that would cause multiple rows in the
relevant membership table with the exact same data; I even saw a few cases of
that in the wild on a dev host. The changes in this commit filter out
duplicates, and even handles the case where an Entity is its own parent.

The logic in this commit is not currently thoroughly tested, apart from a tweak
to an existing test (added in this commit) that tries adding the same child to
an OESS::Entity twice.