Forbid entity self-parentage outright #531
I realized this morning that the code I wrote yesterday to do proper accounting
of cases where an entity is its own parent/child wasn't quite correct. Thinking
further, I then questioned where an entity should be allowed to get into that
state in the first place; per chat, the answer is "no". This fixes things so
that an entity never gets saved to the DB with self-parentage.