Prevent removal of a l3vpn when addVRF is called more than once
Consider the case where addVRF is called twice for the the same
L3VPN. The first call will (likely) succeed. The second call will
certainly fail as the units for the l3vpn are already in use on their
respective devices.

Prior to this change, failure to provision any endpoint inside addVRF
triggered the removal of the l3vpn from all nodes. This meant that
simply calling addVRF twice on the same l3vpn resulted in its entire
removal from the network.

This change modifies how we handle provisioning failures.  If there's
a provisioning failure of an endpoint, the network state for that node
will be left alone (preventing removal of any pre-existing units). The
endpoints that were successfully provisioned will continue to be
removed as before.