Fixing l2 connections requiring site-ids and allowing bad endpoints
A change to the way Endpoints are queried from the database resulted
in a new ordering of the Circuit object’s endpoints array. Because of
how l2vpn/l2vpls site-ids are generated, this change in ordering
resulted an expected diff which would have affected a large number of
layer2 circuits.

This same change also had the built in assumption that all existing
Endpoints would respect OESS’s current ACLs, however some Endpoints
are defined using VLANs outside the existing ACL ranges (some are even
defined using interfaces that are not owned by a workgroup). This
resulted in an expected diff that would have removed of Endpoints of
some circuits and changed the circuit type of others.

Both issues were addressed using the same fix. By splitting up the
single query into multiple we are able to maintain the original
endpoint ordering for proper generation of site-ids, and by removing
an INNER JOIN on the interface_acl table we are able to preserve
endpoints existing outside valid OESS ACLs.