Adding network_type attribute to config tag in database.xml
We previously used the `openflow` and `mpls` attributes to determine
which oess services to start. This change introduces the
`network_type` attribute which basically overrides these old
flags. Values `openflow`, `vpn-mpls`, and `evpn-vxlan` are valid. If a
value is not provided or the value is invalid we default to
`vpn-mpls`. A restriction added by this change si that a dual stack
network (openflow and mpls) may no longer be configured.

This change additionally removes `OESS::Database` as a requirement
from the oess init script.