Handle failures of calls to Switch::get_system_info
get_system_info previously returned undef on error. This caused the
result handler's error check to be bypassed and as a result cleared
the node's loopback address from the database. get_system_info was
modified to properly return an error on failure; The loopback should
now only be cleared if it's unable to be found on the device.
Fixes #829