Add support for multiple ExpressRouteCrossConnections per Connection
The key to this change lies in the AzurePeeringConfig object. This
object is responsible for generation of the prefixes sent to Azure
upon provisioning.

Prefix selection starts with 192.168.100.248/30 for IPv4 and
3FFE:FFFF:0:CD30::0/126 for IPv6. The first Azure Endpoint will use
these prefixes. Every Endpoint after that will use the following
prefixes. For example, the 2nd Endpoint will use 192.168.100.252/30
and 3FFE:FFFF:0:CD30::4/126. The 3rd Endpoint will use
192.168.101.0/30 and 3FFE:FFFF:0:CD30::8/126.

As provisioning with IPv6 using the API is disabled, we also have
disabled IPv6 address selection. To use an IPv6 peering, it must be
set via the Azure Portal by the Customer.

Fixes #1142