add logic to allow edits of azure conns
If v6 peerings are added to an azure conn via oess an error should
be thrown. However users can still add v6 peerings via the azure
portal. When this happens the syncer picks up the addrs and adds
them to oess. Because we still want users to be able to modify
endpoints on these conns, we ignore any pre-existing v6 peerings
when doing connection validation. New v6 peerings will generate
an error.