validate bw set on oess ep matches azure bw
To support bw approval workflow for azure endpoints along with
using the interface-speed config to load bw options in the UI: we
ask that, moving forward, users specify bw when creating azure
endpoints. On provisioning we validate the oess bw matches the
azure bw or we raise an error to ensure consistency.