fix issues relating 1500 mtu on gcp endpoints
Previously the base attributes on an l3 endpoint such as bandwidth
mtu and vlan couldn't be modified. This change enables modification
of mtu on all endpoints. It also enables modification of bandwidth
and tags on non-cloud endpoints. By making these modifications we
allow the gcp syncer to properly configure mtu on gcp endpoints.

This change also updates the l3 connection diff procedure to
support syncing mtu changes to nso.